Move JavaClassUtil.getClassMajorVersion(JavaVersion) to JavaVersion
